Add Google Analytics v3
v3 for backwards compat. We should be logging out exceptions (both
caught and uncaught), screen views, and some key usage metrics.

The dispatch interval is left at the default (30 minutes). I'm not
exactly sure how that works if the app isn't in the foreground. I think
it will go. I guess we'll wait and see.