Fixes for signed build
Fragments can only have a default constructor, switch to newInstance
paradigm.

It's possible that the current tone is a RingPack tone not in our list
of Packs, so avoid NullPointerException if that's the case.

When disabling, why not delete all instances of ourselves in the
MediaStore that we can find?