Fix crash if last tone in pack was disabled
Looks like a wrote a bad while loop in RingPackVm's
findNextEnabledTone(). Forgot to reset the search to the start of the
pack if we're at the end. Also avoids an infinite loop by returning
null.

The infinite loop case is if the user has NO tones enabled in a pack. We
won't let them set it in that case. NOTE: they could go back later and
disable all of them, which will crash us. I'm betting that won't happen.